requirements.txt
certifi==2018.4.16
Django==2.0.4
djangorestframework==3.8.2
pipenv==11.10.0
psycopg2==2.7.4
pytz==2018.4
virtualenv==15.2.0
virtualenv-clone==0.3.0
